requirements.txt
supabase==2.8.0
playwright==1.40.0
python-dotenv==1.0.1
aiohttp==3.9.1
aiofiles==23.2.0
asyncio-throttle==1.0.2
requests==2.31.0
httpx==0.27.0
groq==0.4.1
psycopg2==2.9.9 